CanInterface: Make termination an u16 as the kernel has
As the linux kernel stores (and transmits) can termination as an u16, we
should give that out to users as u16 also.
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/linux/can/dev.h?h=v6.9-rc2#n63
Otherwise bad things can happen. It was an u32 previously.

+++ src/nl/mod.rs
@@ -827,8 +827,8 @@ impl CanInterface {
     }
 
     /// Gets the CANbus termination for the interface
-    pub fn termination(&self) -> Result<Option<u32>, NlInfoError> {
-        self.can_param::<u32>(IflaCan::Termination)
+    pub fn termination(&self) -> Result<Option<u16>, NlInfoError> {
+        self.can_param::<u16>(IflaCan::Termination)
     }
 }
